Harden commerce foundation and add recipe/schema coverage
Prepare the existing commerce modules (Taxation, Products, Subscriptions,
Stripe, Pay Later, Payments, Checkout) as a clean, reusable foundation for
future e-commerce modules, without implementing e-commerce itself.

Foundation hardening:
- Migrate money from double to decimal across Checkout/Payments/Subscriptions/
  Stripe abstractions and models for financial-data integrity.
- Move ProductPart from Payments.Core to Products.Core for correct module
  boundaries; add module-boundary tests.
- Preserve refund/tax integrity and harden Stripe currency handling.
- Fix Stripe webhook idempotency: ConfirmAsync now handles the terminal-state
  payment intent (payment_intent_unexpected_state) by returning the authoritative
  state instead of erroring, so invoice.payment_succeeded returns 200.

Recipes / deployment / schema:
- Add custom part schema definitions for ProductPart, SubscriptionPart,
  SubscriptionSummaryPart and TenantOnboardingPart with feature-gated
  (CrestApps.OrchardCore.Recipes) startups and Recipes.Core references.
- Extend PartSchemaDefinitionTests to cover the new commerce part schemas.
- Confirm Taxation already provides recipe steps, deployment sources/steps/
  drivers and schema for all user-addable catalog entities.

Docs:
- Update products.md and subscriptions.md with Recipes and schema sections.
- Update the 3.0.0 changelog.
- Add planning docs under docs/.

Verified: build 0 warnings/0 errors (warnaserror); 2089 tests pass; docs build
clean. Browser-tested Stripe and Pay Later subscription flows, customer
dashboard/receipts, admin management, and negative authorization.

+++ src/Abstractions/CrestApps.OrchardCore.Checkout.Abstractions/CheckoutInvoice.cs
@@ -20,23 +20,23 @@ public sealed class CheckoutInvoice
     /// <summary>
     /// The one-time amount due now (setup fees, one-off goods), before tax.
     /// </summary>
-    public double? InitialPaymentAmount { get; set; }
+    public decimal? InitialPaymentAmount { get; set; }
 
     /// <summary>
     /// The amount of the first recurring cycle charged now, before tax.
     /// </summary>
-    public double? FirstRecurringPaymentAmount { get; set; }
+    public decimal? FirstRecurringPaymentAmount { get; set; }
 
     /// <summary>
     /// The total amount collected now, before tax.
     /// </summary>
-    public double DueNow { get; set; }
+    public decimal DueNow { get; set; }
 
     /// <summary>
     /// The tax charged on the amount due now, determined by the taxation framework. This is <c>0</c> when
     /// the Taxation feature is disabled.
     /// </summary>
-    public double TaxAmount { get; set; }
+    public decimal TaxAmount { get; set; }
 
     /// <summary>
     /// The detailed tax lines that explain <see cref="TaxAmount"/>. Each jurisdiction/tax is preserved.
@@ -62,13 +62,13 @@ public sealed class CheckoutInvoice
     /// <summary>
     /// The grand total collected now, including tax.
     /// </summary>
-    public double GrandTotal { get; set; }
+    public decimal GrandTotal { get; set; }
 
     /// <summary>
     /// The per-interval subtotals for the recurring obligations, before tax.
     /// </summary>
     [JsonConverter(typeof(BillingDurationKeyDictionaryJsonConverter))]
-    public Dictionary<BillingDurationKey, double> Subtotals { get; set; }
+    public Dictionary<BillingDurationKey, decimal> Subtotals { get; set; }
 
     /// <summary>
     /// Every priced line on the invoice.
