+++ src/Abstractions/CrestApps.OrchardCore.Payments.Abstractions/Money.cs
@@ -2,14 +2,13 @@ namespace CrestApps.OrchardCore.Payments;
 
 /// <summary>
 /// Provides safe, currency-aware comparison and rounding for monetary amounts carried as
-/// <see cref="double"/> values.
+/// <see cref="decimal"/> values.
 ///
-/// Comparing amounts with the default <c>==</c>/<c>!=</c> operators is unsafe because binary floating
-/// point cannot represent most decimal fractions exactly (for example <c>19.99 + 10.00</c> is not
-/// guaranteed to equal <c>29.99</c>). That imprecision can cause a valid payment to be rejected, or two
-/// different amounts to be treated as equal. All comparisons are therefore performed after normalizing
-/// the amounts to whole minor units for the relevant currency, which is the smallest unit any supported
-/// gateway settles in, so the result is deterministic and gateway aligned.
+/// Money is represented as <see cref="decimal"/>, the authoritative type for financial values, so amounts
+/// are exact at the currency's own scale. Comparisons are still performed after normalizing the amounts to
+/// whole minor units for the relevant currency — the smallest unit any supported gateway settles in — so a
+/// value expressed at a finer precision than the currency supports can never make two settlement-equal
+/// amounts compare as different, and the result stays deterministic and gateway aligned.
 /// </summary>
 public static class Money
 {
@@ -18,7 +17,7 @@ public static class Money
     /// </summary>
     /// <param name="amount">The amount expressed in major units.</param>
     /// <param name="currency">The ISO-4217 currency code used to determine precision. Defaults to two decimals when unknown.</param>
-    public static double Round(double amount, string currency = null)
+    public static decimal Round(decimal amount, string currency = null)
         => Math.Round(amount, CurrencyScale.GetDecimalPlaces(currency), MidpointRounding.AwayFromZero);
 
     /// <summary>
@@ -26,8 +25,8 @@ public static class Money
     /// </summary>
     /// <param name="amount">The amount expressed in major units.</param>
     /// <param name="currency">The ISO-4217 currency code used to determine precision.</param>
-    public static long ToMinorUnits(double amount, string currency = null)
-        => CurrencyScale.ToMinorUnits((decimal)amount, currency);
+    public static long ToMinorUnits(decimal amount, string currency = null)
+        => CurrencyScale.ToMinorUnits(amount, currency);
 
     /// <summary>
     /// Determines whether two monetary amounts are equal once normalized to whole minor units for the currency.
@@ -35,7 +34,7 @@ public static class Money
     /// <param name="left">The first amount expressed in major units.</param>
     /// <param name="right">The second amount expressed in major units.</param>
     /// <param name="currency">The ISO-4217 currency code used to determine precision.</param>
-    public static bool AreEqual(double left, double right, string currency = null)
+    public static bool AreEqual(decimal left, decimal right, string currency = null)
         => ToMinorUnits(left, currency) == ToMinorUnits(right, currency);
 
     /// <summary>
@@ -45,7 +44,7 @@ public static class Money
     /// <param name="left">The first amount expressed in major units.</param>
     /// <param name="right">The second amount expressed in major units.</param>
     /// <param name="currency">The ISO-4217 currency code used to determine precision.</param>
-    public static bool AreEqual(double? left, double? right, string currency = null)
+    public static bool AreEqual(decimal? left, decimal? right, string currency = null)
         => left.HasValue && right.HasValue && AreEqual(left.Value, right.Value, currency);
 
     /// <summary>
@@ -55,6 +54,6 @@ public static class Money
     /// <param name="amount">The amount to test, expressed in major units.</param>
     /// <param name="threshold">The threshold to compare against, expressed in major units.</param>
     /// <param name="currency">The ISO-4217 currency code used to determine precision.</param>
-    public static bool IsGreaterThan(double amount, double threshold, string currency = null)
+    public static bool IsGreaterThan(decimal amount, decimal threshold, string currency = null)
         => ToMinorUnits(amount, currency) > ToMinorUnits(threshold, currency);
 }
