+++ src/Core/CrestApps.OrchardCore.Subscriptions.Core/Models/SubscriptionPart.cs
@@ -16,7 +16,7 @@ public sealed class SubscriptionPart : ContentPart
     /// <summary>
     /// Gets or sets the one-time amount charged when the subscription starts.
     /// </summary>
-    public double? InitialAmount { get; set; }
+    public decimal? InitialAmount { get; set; }
 
     /// <summary>
     /// Gets or sets the number of <see cref="DurationType"/> units in one billing cycle.
